Fix unittest issue
