+++ src/Fhir.Anonymizer.Core.UnitTests/Visitors/AnonymizationVisitorTests.cs
@@ -44,6 +44,30 @@ namespace Fhir.Anonymizer.Core.UnitTests.Visitors
         }
 
         [Fact]
+        public void GivenACryptoHashRule_WhenProcess_NodeShouldBeHashed()
+        {
+            AnonymizationFhirPathRule[] rules = new AnonymizationFhirPathRule[]
+            {
+                new AnonymizationFhirPathRule("Patient.address", "address", "Patient", "cryptoHash", AnonymizerRuleType.FhirPathRule, "Patient.address"),
+            };
+
+            AnonymizationVisitor visitor = new AnonymizationVisitor(rules, CreateTestProcessors());
+
+            var patient = CreateTestPatient();
+            var patientNode = ElementNode.FromElement(patient.ToTypedElement());
+            patientNode.Accept(visitor);
+            patientNode.RemoveNullChildren();
+
+            var patientAddress = patientNode.Select("Patient.address[0].city").FirstOrDefault();
+            Assert.Equal("c4321653de997f3029d2efa38dd4baa6c9c2f6bd67b8a52be789f157f8b286ce", patientAddress.Value.ToString());
+
+            patient = patientNode.ToPoco<Patient>();
+            Assert.Single(patient.Meta.Security);
+            Assert.Contains(SecurityLabels.CRYTOHASH.Code, patient.Meta.Security.Select(s => s.Code));
+        }
+
+
+        [Fact]
         public void GivenAPatientWithOnlyId_WhenProcess_NodeShouldBeRedact()
         {
             AnonymizationFhirPathRule[] rules = new AnonymizationFhirPathRule[]
@@ -80,7 +104,7 @@ namespace Fhir.Anonymizer.Core.UnitTests.Visitors
 
             var patient = CreateTestPatient();
             var patientNode = ElementNode.FromElement(patient.ToTypedElement());
-            patientNode.Accept(visitor); 
+            patientNode.Accept(visitor);
             string patientCity = patientNode.Select("Patient.address[0].city").First().Value.ToString();
             string patientCountry = patientNode.Select("Patient.address[0].country").First().Value.ToString();
 
@@ -311,11 +335,13 @@ namespace Fhir.Anonymizer.Core.UnitTests.Visitors
             KeepProcessor keepProcessor = new KeepProcessor();
             RedactProcessor redactProcessor = new RedactProcessor(false, false, false, null);
             DateShiftProcessor dateShiftProcessor = new DateShiftProcessor("123", "123", false);
+            CryptoHashProcessor cryptoHashProcessor = new CryptoHashProcessor("123");
             Dictionary<string, IAnonymizerProcessor> processors = new Dictionary<string, IAnonymizerProcessor>()
             {
                 { "KEEP", keepProcessor},
                 { "REDACT", redactProcessor},
-                { "DATESHIFT", dateShiftProcessor}
+                { "DATESHIFT", dateShiftProcessor},
+                { "CRYPTOHASH", cryptoHashProcessor}
             };
 
             return processors;
